Move to prediction head adding methods to base mixin
