+++ src/adapters/heads/__init__.py
@@ -2,3 +2,4 @@
 from .base import *
 from .dependency_parsing import *
 from .language_modeling import BertStyleMaskedLMHead, CausalLMHead, Seq2SeqLMHead
+from .model_mixin import ModelWithFlexibleHeadsAdaptersMixin
